fix a few tests
