--- examples/experimental/test/end_to_end/http_server/sqlite/src/server/index.ts
@@ -1,24 +1,0 @@
-import express from 'express';
-import { Server as NodeServer } from 'http';
-
-import { getRouter as getRouterPosts } from '../entities/posts/router';
-import { getRouter as getRouterUsers } from '../entities/users/router';
-
-export function initServer(): NodeServer {
-    let app = express();
-
-    app.use(express.json());
-
-    app.use('/users', getRouterUsers());
-    app.use('/posts', getRouterPosts());
-
-    app.get('/init-called', (_req, res) => {
-        res.json(globalThis._azleInitCalled);
-    });
-
-    app.get('/post-upgrade-called', (_req, res) => {
-        res.json(globalThis._azlePostUpgradeCalled);
-    });
-
-    return app.listen();
-}
